review: int radio status (no narrowing), precise RangeMath equivalence wording
Copilot x3 (2 distinct), both accepted: (1) dashboardServiceRadioStatus
returns int, forwarding WiFi.status() without narrowing (all defined
statuses fit either way — printed decimal unchanged; the redundant call-
site cast dropped). (2) RangeMath.h comment now claims value-result
equivalence for side-effect-free arguments (every call site) instead of
macro-level semantics. 27/27 suites green, flash/RAM unchanged.

+++ sketches/rc_test/rc_test.ino
@@ -708,7 +708,7 @@ void wifiDebug(uint32_t nowUs) {
   // status/counters, two-digit uppercase hex + trailing space per byte).
   char line[96];
   snprintf(line, sizeof(line), "# WIFI up=%d status=%d clients_seq=%lu",
-           wifiUp ? 1 : 0, (int)dashboardServiceRadioStatus(), (unsigned long)wifiSeq);
+           wifiUp ? 1 : 0, dashboardServiceRadioStatus(), (unsigned long)wifiSeq);
   debugConsolePrintLine(line);
 
   // X.BUS RX byte-level diagnostics — tells us whether D0 sees anything at all.
