+++ sketches/rc_test/src/infrastructure/network/WifiService.cpp
@@ -25,7 +25,7 @@ WiFiClient pageClient;
 const char *pagePointer = nullptr;
 size_t     pageRemaining = 0;
 
-uint8_t dashboardServiceRadioStatus() { return WiFi.status(); }
+int dashboardServiceRadioStatus() { return WiFi.status(); }
 
 bool dashboardServiceInitialize(const char* ssid, const char* password,
                                 uint8_t apChannel, const char* pageHtml) {
